// app/(employees)/shifts/types.ts
  export interface itemListProps {
    // key?: string;
    item: string;
}

export interface ShiftSearchParams {
  shiftType: string;
  date: string,
  time: string
  hourlyRate: string,
  numberOfHours: string,
  numberOfOpenings: string,
}